Add macro, testing, and reorganize into subcrates
