+++ .gitignore
@@ -1,3 +1,6 @@
 /target
 **/target
-examples/**/Cargo.lock
+/Cargo.lock
+**/Cargo.lock
+.envrc
+rust-toolchain.toml
