+++ .vscode/launch.json
@@ -104,6 +104,6 @@
             "presentation": {
                 "group": "Windows"
             }
-        },
+        }
     ]
 }
