+++ crates/unbug_bevy/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "unbug_bevy"
-documentation = "https://docs.rs/unbug_bevy"
+documentation = "https://docs.rs/unbug"
 description = "Bevy integration for the unbug debug assertion library"
 version.workspace = true
 edition.workspace = true
@@ -12,6 +12,4 @@ categories.workspace = true
 
 [dependencies]
 bevy_ecs.workspace = true
-unbug_assert.workspace = true
-unbug_macro.workspace = true
