+++ examples/testing/.vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+    "recommendations": [
+        "rust-lang.rust-analyzer"
+    ]
+}
