+++ examples/testing/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "unbug_testing_example"
+version = "1.0.0"
+edition.workspace = true
+publish = false
+
+[dependencies.unbug]
+path = "../../"
+default-features = false
+features = [
+    "check_debugger",
+    "cache_debugger",
+]
+
+# All of the following is necessary to activate unbug's panics during tests
+
+# HACK: Cargo libraries do not get --cfg test passed during tests
+# see: https://github.com/rust-lang/cargo/issues/2911
+[dev-dependencies]
+# replace `unbug_basic_example` with your package name
+unbug_testing_example = { path = ".", features = ["testing"] }
+
+[features]
+default = []
+testing = ["unbug/testing"]
+# this feature is needed to avoid "unreachable" warnings from rust-analyzer
+rust_analyzer = ["unbug/rust_analyzer"]
